==> Notebook/1-Dataset_Extracting.ipynb <==
**This notebook is part of a larger project aimed at performing a comparative analysis of hotel reviews. In this section, we focus on processing and analyzing user reviews from a dataset containing JSON Lines files. We begin by reading the JSON data into a Pandas DataFrame, followed by cleaning and transforming the data for further analysis. The primary objectives include extracting and structuring review data, filtering for specific regions (New York City), and analyzing various aspects of customer feedback such as service, cleanliness, and overall ratings. Additionally, we clean the review text to facilitate text analysis and sentiment evaluation in later stages of the project.**

import json
import pandas as pd
import re

file_path = "./data/review.txt"

# Read the JSON Lines file with Pandas
df = pd.read_json(file_path, lines=True)

df.shape

off = pd.read_json("./data/offering.txt",lines=True)

off

df.head()

df = pd.merge(df, off, left_on='offering_id', right_on='id', how = 'left')

df.head()

df['date'].nunique()

# Function to check if the address contains 'NY' region
def is_nyc_area(address):
    if isinstance(address, dict):
        return address.get('region') == 'NY'
    return False

# Filter the DataFrame
df_nyc = df[df['address'].apply(is_nyc_area)]

# Display the first few rows of the filtered DataFrame
df_nyc.head()


print(type(df_nyc))

df_nyc['name'].nunique()

df_nyc['name'].unique()

# Function to extract all unique keys in the ratings field
def get_unique_keys(column):
    unique_keys = set()
    for entry in column:
        if isinstance(entry, dict):
            unique_keys.update(entry.keys())
    return unique_keys

# Extract unique keys from the ratings column
unique_rating_keys = get_unique_keys(df_nyc['ratings'])

unique_rating_keys

# Function to extract ratings from JSON-like string or a dictionary
def extract_ratings(rating):
    if isinstance(rating, str):
        try:
            return json.loads(rating.replace("'", "\""))
        except json.JSONDecodeError:
            return None
    elif isinstance(rating, dict):
        return rating
    else:
        return None

# Apply the function to the ratings column
ratings_extracted = df_nyc['ratings'].apply(extract_ratings)

# Define the function to extract a specific rating category from the dictionary
def extract_rating_category(ratings_dict, category):
    return ratings_dict.get(category, None) if ratings_dict else None

# List of categories to extract
categories = ['service', 'cleanliness', 'overall', 'location', 'rooms', 'sleep_quality', 'value', 'business_service_(e_g_internet_access)', 'check_in_front_desk']


df_nyc = df_nyc.copy()

# Apply the function to each category and create new columns
for category in categories:
    df_nyc[f'{category}_rating'] = ratings_extracted.apply(lambda x: extract_rating_category(x, category))

df_nyc.columns

df_nyc.drop(columns=['ratings', 'author', 'date_stayed', 'offering_id', 'id_x', 'via_mobile', 'region_id', 'url', 'phone', 'details', 'address', 'id_y'], inplace=True)


df_nyc.drop(columns=['num_helpful_votes', 'type', 'hotel_class'])

def clean_text(text):
    """
    Function to clean text by removing unwanted characters, converting to lowercase,
    and removing extra whitespace.

    Parameters:
    text (str): The text to be cleaned.

    Returns:
    str: The cleaned text.
    """
    # Convert text to lowercase
    text = text.lower()
    
    # Remove URLs
    text = re.sub(r'http\S+|www\S+|https\S+', '', text, flags=re.MULTILINE)
    
    # Remove HTML tags
    text = re.sub(r'<.*?>', '', text)
    
    # Remove punctuation and numbers
    text = re.sub(r'[^a-z\s]', '', text)
    
    # Remove extra whitespace
    text = re.sub(r'\s+', ' ', text).strip()
    
    return text

# Apply the clean_text function to each row in the 'text' column
df_nyc['cleaned_text'] = df_nyc['text'].apply(clean_text)

df_nyc.to_csv('./data/nyc_hotels.csv', index = False)
