==> Notebook/2_Detect_Language.ipynb <==
**In this notebook, we focus on identifying and isolating English-language reviews from a larger dataset of hotel reviews. The process begins by loading the review data and applying a language detection algorithm to each review text. Reviews identified as English are then filtered for further analysis. Additionally, the notebook includes basic text cleaning steps, such as removing HTML tags, punctuation, and converting text to lowercase, to prepare the data for subsequent analysis. The cleaned and filtered English reviews are then saved to a new CSV file for use in later stages of the project.**

from langdetect import detect
import pandas as pd
import re
import string

df = pd.read_csv('./data/nyc_hotels.csv')

# Function to detect language
def detect_language(text):
    if pd.isna(text):
        return "unknown"
    try:
        return detect(text)
    except:
        return "unknown"

# Apply the language detection function to the review text
df['language'] = df['text'].apply(detect_language)

reviews_df = df[df['language'] == 'en']

print(f'All languages shape {df.shape}')
print(f'En languages shape {reviews_df.shape}')

# Data Cleaning

import pandas as pd
import re
import string

# Function to clean text
def basic_clean_text(text):
    # Remove HTML tags
    text = re.sub(r'<.*?>', '', text)
    
    # Remove punctuation
    text = text.translate(str.maketrans('', '', string.punctuation))
    
    # Convert to lowercase
    text = text.lower()
    
    return text

# Apply the basic_clean_text function to the 'text' column
reviews_df.loc[:, 'cleaned_text'] = reviews_df['text'].apply(basic_clean_text)

# Display the first few rows of the cleaned dataset
reviews_df[['text', 'cleaned_text']].head()

reviews_df.to_csv('./data/eng_reviews.csv', index = False)
