==> Notebook/4-Transformer.ipynb <==
**In this notebook, we leverage a transformer-based model to perform sentiment analysis on hotel reviews from New York City. The process includes the following steps:**

Data Preparation: The notebook begins by loading a dataset that includes pre-processed hotel reviews with previously identified topics.

Sentiment Analysis Setup: We use the Hugging Face transformers library to set up a sentiment analysis pipeline. The DistilBERT model, fine-tuned on the SST-2 dataset for sentiment classification, is chosen for this task.

Sentiment Prediction: The model is applied to the text data, predicting sentiment labels (e.g., positive or negative) along with confidence scores. These results are added back to the original dataset for further analysis.

Results Integration: The dataset is updated with the sentiment labels and scores, providing a comprehensive view of customer sentiments associated with each review. This enriched data can be used for deeper analysis, including correlating sentiments with specific topics or hotel attributes.

This notebook concludes by saving the sentiment-enriched dataset, setting the stage for further exploration and insights into customer satisfaction and hotel performance.

import pandas as pd
import re
import string
import torch
from transformers import pipeline, AutoTokenizer, AutoModelForSequenceClassification
from datasets import Dataset, DatasetDict

reviews_df = pd.read_csv('./data/df_with_topics.csv')

# Convert the DataFrame to a Hugging Face Dataset
dataset = Dataset.from_pandas(reviews_df[['cleaned_text']])
# Specify the model and tokenizer
model_name = "distilbert-base-uncased-finetuned-sst-2-english"
tokenizer = AutoTokenizer.from_pretrained(model_name)

# For GPU usage
device = 0 if torch.cuda.is_available() else -1
# Load the sentiment analysis pipeline
sentiment_analysis = pipeline("sentiment-analysis", model=model_name, tokenizer=tokenizer, device=device)


# Define a function to get the sentiment
def get_sentiment(batch):
    # Truncate text to the maximum length of the model
    truncated_texts = [text[:tokenizer.model_max_length] for text in batch['cleaned_text']]
    sentiments = sentiment_analysis(truncated_texts)
    return {
        'sentiment': [s['label'] for s in sentiments],
        'score': [s['score'] for s in sentiments]
    }

# Apply the sentiment analysis in batches
result = dataset.map(lambda batch: get_sentiment(batch), batched=True, batch_size=58)

# Add the sentiment results back to the original DataFrame
reviews_df['sentiment'] = result['sentiment']
reviews_df['score'] = result['score']

# Display the first few rows of the dataset with sentiment labels and scores
print(reviews_df[['cleaned_text', 'sentiment', 'score']].head())

reviews_df

reviews_df.to_csv('./data/reviews.csv', index= False)
